requirements.txt
Flask==3.1.1
numpy==1.25.1
python-dotenv==1.1.1
pytz==2025.2
Requests==2.32.4
river==0.22.0
solders==0.26.0
feedparser==6.0.10
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch>=2.0.0
gymnasium>=0.29.0
pandas>=1.5.0
shap>=0.42.0
